now deletes an activations in the manageprofiles page refresh the page
see title. now deletes an activations in the manageprofiles page refresh the page. Basically this. Not sure this is done in the most effective ways, was seeing that there might be a way to use keys to do it, but it is quick and not too hard so thats pretty cool